A little more project directory reshuffling to fix some non-relative links in the workspace.

+++ Classes/TGRESTInMemoryStore.m
@@ -258,7 +258,7 @@
         objectCount = objectCount + resourceArray.count;
     }
     
-    return [NSString stringWithFormat:@"%@ with %lu resources and %lu objects", [[self class] description], self.inMemoryDatastore.allKeys.count, objectCount];
+    return [NSString stringWithFormat:@"%@ with %lu resources and %lu objects", [[self class] description], (unsigned long)self.inMemoryDatastore.allKeys.count, objectCount];
 }
 
 @end
